Fix Loader::$essentialspe field
You will not find any bugs with it because that problem doesn't affect at the moment, but they won't like it. If you want to use that field, always use `$loader->essentialspe->get()`, but use `$loader->essentialspe->valid()` first to check if EssentialsPE is unloaded.

I'm sorry. Although you refuse to merge the pull request, I have to do it because it is wrong.